Add bytes_view
This presents a view of a range as a sequence of bytes, as you
might have guessed. The idea is that it makes it easier to pass
to an ostream which writes out bytes to disk.

I still need to work out the exact constraints on the range value
type. SimpleType? POD? Fundemental type?