Add ostreambuf_iterator.hpp
So Range-V3 doesn't thing that std::ostreambuf_iterator is a real
output iterator, I think because it's not DefaultConstructible
(and therefore not Semiregular and therefore not WeaklyIncrementable).

This seems a bit silly, but never mind. Here's an ostreambuf_iterator
that it does like. The code is based on Range-V3's ostream_iterator,
which presumably exists for the same reason...