Add default constructor to concat_range()
This is the (poorly named) view that gets used when adding a BOM to an
InputRange. It was playing up because it wasn’t matching the View
requirements. It turns out that’s because it was missing a default
constructor, which took me ages to work out.

C++ error messages suck.