Rewrite "concat_range" as "bom_concat_view"
The new version is now a proper view_adaptor, rather than trying to be
fancy by implementing view_interface and not really working.

While it’s still a private class and not intended to be generally used,
it also now has a much better name.

Plus it works, so that’s nice.